Changed type of varaible N_S from double to long.

+++ release1.1/src/plugins/coremodules/Classpbsm.h
@@ -43,7 +43,7 @@ const double *Drift_offset{ NULL };
 const double *Drift_slope{ NULL };
 const double *Subl_offset{ NULL };
 const double *Subl_slope{ NULL };
-const double *N_S{ NULL };
+const long *N_S{ NULL };
 const long *inhibit_evap{ NULL };
 const long *inhibit_bs{ NULL };
 const long *inhibit_subl{ NULL };
